replaced "GNU Linux" with "GNU/Linux"
When you write "GNU Something", that means "Something" is a part of the GNU project. Although Linux is a free software, but it's not a part of the GNU project. For meaning the "GNU and Linux" operating system, write, "GNU/Linux". For more info on it, https://www.gnu.org/gnu/gnu-linux-faq.html#whyslash